% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_RXProgID.R
\name{add_RXProgID}
\alias{add_RXProgID}
\title{add_RXProgID}
\usage{
add_RXProgID(ItemMeta1, ServiceType, ProgName, ProgDescription,
  db.settings)
}
\arguments{
\item{ItemMeta1}{User group aka where to file this program under}

\item{ServiceType}{Typically Community or Governance.}

\item{ProgName}{The program name}

\item{ProgDescription}{The Program description}

\item{db.settings}{Database user, host, pw}
}
\description{
Add a program to the RX_Master lsit
}
\examples{
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName')
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName',ItemMeta1='New user group')
Need to specify the RX_ProgID and then reference a column for updating. You may specify any combinations of ItemMeta1, ServiceType, ProgName, or ProgDEscription for updating
}
