% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_RXProgID.R
\name{edit_RXProgID}
\alias{edit_RXProgID}
\title{edit_RXProgID}
\usage{
edit_RXProgID(RX_ProgID, ItemMeta1 = NULL, ServiceType = NULL,
  ProgName = NULL, ProgDescription = NULL, db.settings)
}
\arguments{
\item{RX_ProgID}{The ID of the program to update}

\item{ItemMeta1}{User group aka where to file this program under}

\item{ServiceType}{Typically Community or Governance.}

\item{ProgName}{The program name}

\item{ProgDescription}{The Program description}
}
\description{
Update an RX master program by ID
}
\examples{
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName')
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName',ItemMeta1='New user group')
Need to specify the RX_ProgID and then reference a column for updating. You may specify any combinations of ItemMeta1, ServiceType, ProgName, or ProgDEscription for updating
}
