% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RXProgID_Info.R
\name{get_RXProgID_Info}
\alias{get_RXProgID_Info}
\title{get_RXProgID_Info}
\usage{
get_RXProgID_Info(RX_ProgID)
}
\arguments{
\item{RX_ProgID}{The ID of the program info to get. Set to NULL to pull all programs. Provide a vector of indexes to pull a subset.}
}
\description{
Update an RX master program by ID
}
\examples{
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName')
edit_RXProgID(RX_ProgID=1,ProgName='Updated ProgName',ItemMeta1='New user group')
Need to specify the RX_ProgID and then reference a column for updating. You may specify any combinations of ItemMeta1, ServiceType, ProgName, or ProgDEscription for updating
}
