% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/globalsearch.R
\name{globalsearch}
\alias{globalsearch}
\title{globalsearch}
\usage{
globalsearch(RXProgID, DatabaseNames = NULL, keyword = NULL)
}
\arguments{
\item{RXProgID}{names of the PBB databases to summarrize, set to NULL for all databases. Will exclude test databases}

\item{DatabaseNames}{names of the PBB databases to summarrize, set to NULL for all databases. Will exclude test databases}

\item{keyword}{keyword to search program names and descriptions for}
}
\description{
This function is the same as summarize_RXProgID_Data, but with a keyword parameter. We could consolidate into one function, but what to call it.
}
\examples{
data<-globalsearch(RXProgID=NULL,DatabaseNames=NULL,keyword='Fireworks')
}

