% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_RXProgIDs.R
\name{summarize_RXProgIDs}
\alias{summarize_RXProgIDs}
\title{summarize_RXProgIDs}
\usage{
summarize_RXProgIDs(DatabaseNames = NULL, db.settings)
}
\arguments{
\item{DatabaseNames}{names of the PBB databases to summarrize, set to NULL for all databases. Will exclude test databases}

\item{db.settings}{Database user, host, pw}
}
\description{
Creates two summary dataframes. One is a summary by Program - how often each program is used The second a sumary by Org - how many mapped programs per org.
}
\examples{
data<-summarize_RXProgIDs(DatabaseNames)
write.csv(data$RX_ProgInfo,'RX_Proginfo.csv')
write.csv(data$Org_Program_Data,'Org_Program_Data.csv')
}
